% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lintr.R
\name{lintr}
\alias{lintr}
\title{Help with dplyr command construction}
\usage{
lintr(Input)
}
\arguments{
\item{Input}{A data frame or piped input from a chain of \code{dplyr} commands.}
}
\description{
\code{lintr()} opens an interactive (shiny) app that lets you edit
interactively a
potentially complex \code{dplyr} chain of commands. It provides documentation
and argument vetting.
}
\details{
On return, \code{lintr()} opens and editor to display and run the command.  Often,
the user will cut and paste this into a chunk in an Rmd document.

The name "lintr" refers to the famous \code{lint} program distributed with UNIX
that checks C program files against common mistakes.
}
\examples{
\dontrun{require(mosaicData)
lintr( KidsFeet \%>\% select( sex, length, width, biggerfoot) )
}
}

