% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mPlots.R
\name{mUSMap}
\alias{mBar,}
\alias{mDistribution}
\alias{mScatter,}
\alias{mUSMap}
\alias{mUSMap,}
\alias{mWorldMap,}
\title{Interactive interface to ggplot2}
\usage{
mUSMap(data = NULL, key = NULL, fill = NULL, ...)

mDistribution(data, format = "histogram", default = format,
  system = "ggplot2", show = FALSE, title = "", ...)
}
\arguments{
\item{data}{Dataframe containing the variables that might be used in the plot.}

\item{key}{name of variable holding the state or country ID}

\item{fill}{name of variable to use for the cholopleth map}
}
\value{
Nothing.  Just for plotting side effects.
}
\description{
Provides a menu selection system (via \code{manipulate}) so that the variables
for different aspects of a plot can be selected interactively.  The \code{ggplot2}
command for generating the plot currently being displayed is copied to the console,
whence it can be copied to a document for later direct, non-interactive use.
}

