% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mWrangle.R
\name{mWrangle}
\alias{mWrangle}
\title{Help with dplyr command construction}
\usage{
mWrangle(Input)
}
\arguments{
\item{Input}{A data frame or piped input from a chain of \code{dplyr} commands.}
}
\description{
\code{mWrangle()} opens an interactive (shiny) app that lets you edit
interactively a
potentially complex \code{dplyr} chain of commands. It provides documentation
and argument vetting.
}
\details{
On return, \code{mWrangle()} opens and editor to display and run the command.  Often,
the user will cut and paste this into a chunk in an Rmd document.

The "m" in "mWrangle" follows the convention that a leading "m"
indicates an interactive command.
}
\examples{
\dontrun{require(mosaicData)
mWrangle( KidsFeet \%>\% select( sex, length, width, biggerfoot) )
}
}

