% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mPlots.R
\name{mScatter}
\alias{mBar}
\alias{mBar,}
\alias{mPlots,}
\alias{mScatter}
\alias{mScatter,}
\alias{mWorldMap}
\title{Interactive interface to ggplot2}
\usage{
mScatter(dat)

mBar(dat)

mWorldMap(dat, countryVar = "Country")
}
\arguments{
\item{dat}{Dataframe containing the variables that might be used in the plot.}

\item{countryVar}{name of variable with the country names --- default "Country"}
}
\value{
Nothing.  Just for side effects.
}
\description{
Provides a menu selection system (via \code{manipulate}) so that the variables
for different aspects of a plot can be selected interactively.  The \code{ggplot2}
command for generating the plot currently being displayed is copied to the console,
whence it can be copied to a document for later direct, non-interactive use.
}

